+++ src/RestApi/CustomScript/ICustomScript.ts
@@ -6,5 +6,5 @@ export interface ICustomScript {
 	dangerLevel: string;
 }
 
-export type ICustomScriptCreatable = Pick<ICustomScript, 'name' | 'title' | 'description' | 'dangerLevel'>;
+export type ICustomScriptCreatable = Pick<ICustomScript, 'name' | 'title' | 'description' | 'dangerLevel'> & { organizationUid?: string };
 export type ICustomScriptUpdatable = Partial<Pick<ICustomScript, 'name' | 'title' | 'description' | 'dangerLevel'>>;